requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
langgraph==0.4.7
langchain==0.3.25
databricks-langchain==0.5.1
unitycatalog-langchain[databricks]==0.2.0
unitycatalog-ai[databricks]==0.3.1
langgraph-checkpoint-postgres==2.0.21
duckduckgo-search==8.0.2
databricks-agents==0.22.1
psycopg[binary,pool]==3.2.9
databricks-sdk==0.54.0
langgraph-reflection==0.0.1
openevals==0.1.0
mlflow==2.22.0
pydantic==2.11.5
loguru==0.7.3
rich>=13.7.0
databricks-connect==16.3.3
python-dotenv>=1.0.0